(vimrc) Remove rogue line that ruins MySmartResize()
The function's output would mention that a particular pattern wouldn't match.
I can't begin to guess what I was thinking when I left an ill-formed line in the
function that vim tried to parse as a do-nothing `:global` command. I vaguely
recall it not working without it, but if I had to guess, I simply forgot to
re-run my vimrc and misread the error messages. Idunno.