[hgrc] Fix hg-git extension for Mercurial
The hg-git extension for Mercurial allows a local mercurial repository
to treat a remote git repository like a mercurial repository. This is
particularly important as bitbucket prepares to drop support for
Mercurial, deleting all still-present Mercurial repositories at the end
(beginning?) of June.

Fixing support for this extension requires manually installing the
extension, whereas before, I had used tortoisehg as a sort of collection
of useful Mercurial tools and a history graph viewer.

Only the two following commands are required to complete the fix for a
local machine:

```
pip install dulwich

hg clone https://foss.heptapod.net/mercurial/hg-git ~/code/hg-git
```

Sans any clever scripting, cloning into that specific location is
required, since the .hgrc settings call for a static, actual path. I may
do some elaborate scripting in the future, but this *is* a repository of
personal settings, and I *actually use* that path, so it may be overkill.

It's also possible I'll change the clone path for hg-git to something
that has a stronger semantic for "actively being used". the `~/code`
folder is mostly used for personal projects and interesting code bases
found on the internet. In Windows, I have the folder
`'C:\Program Files (Manually Installed)'`, aping the location of the
automatic `'C:\Program Files'` folder, but its root location is
innappropriate for a cross-platform folder.